+++ src/parse.rs
@@ -3,7 +3,7 @@ mod current;
 mod history;
 mod proces;
 
-pub use ansi::{Ansi, AnsiStr};
+pub use ansi::{Ansi, AnsiStr, Theme};
 pub use current::{get_buildlog, get_emerge, get_pretend, get_resume, Pkg};
 pub use history::{get_hist, Hist, MergeStep};
 #[cfg(test)]